requirements.txt
telethon
tqdm
numpy
pillow
python-pptx
pywin32
pycairo
natsort